move client options into standalone generator
